requirements.txt
streamlit
torch
torchvision
pandas
pillow